+++ discopy/__init__.py
@@ -33,6 +33,7 @@ from discopy import (
     grammar,
     drawing,
     utils,
+    testing,
     config,
     messages,
 )